pyproject.toml
[project]
name = "restaurant-financial-mcp"
version = "0.1.0"
description = "MCP Server for Chinese Restaurant Financial Analysis - Powered by Claude Intelligence"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Heng Woo", email = "hengwoo@example.com" }
]
license = { text = "MIT" }
keywords = ["mcp", "restaurant", "financial", "analysis", "chinese", "claude", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.15.0",
    "mcp>=1.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "asyncio-mqtt>=0.11.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.2",
    "watchdog>=3.0.0",
    "click>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/HengWoo/fin_report_agent"
Documentation = "https://github.com/HengWoo/fin_report_agent#readme"
Repository = "https://github.com/HengWoo/fin_report_agent"
Issues = "https://github.com/HengWoo/fin_report_agent/issues"

[project.scripts]
restaurant-mcp = "src.mcp_server.cli:main"
restaurant-mcp-server = "src.mcp_server.cli:start_server"
restaurant-mcp-setup = "src.mcp_server.cli:setup_claude"
restaurant-mcp-test = "src.mcp_server.cli:test"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
